requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
future==1.0.0
geocoder==1.38.1
idna==3.7
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pathspec==0.12.1
platformdirs==4.2.2
probableparsing==0.0.1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
ratelim==0.1.6
requests==2.32.3
setuptools==70.1.0
six==1.16.0
tzdata==2024.1
urllib3==2.2.2
usaddress==0.5.10
usaddress-scourgify==0.6.0
yaml-config==0.1.5
